OMA-DM: fix compiling for 64-bit
Fix building for 64-bit by casting to intptr_t when converting
a pointer to an int.
Disable building libdmengine for 64-bit, it puts too many
pointers into UINT32.

+++ engine/dmlib/api/native/dmtErrorDescription.hpp
@@ -188,7 +188,7 @@ public:
   */
   inline BOOLEAN operator==(void* p) const 
   { 
-      return m_nErrorCode == (int)p; 
+      return m_nErrorCode == (intptr_t)p; 
   }
   
  /**
@@ -204,7 +204,7 @@ public:
   */
   inline BOOLEAN operator!=(void* p) const 
   { 
-      return m_nErrorCode != (int)p;
+      return m_nErrorCode != (intptr_t)p;
   }
 
  /**
